Geocoding code
I fixed a couple things - one being that it was adding NA values back to the unique_city list, which I think was causing issues. I also have it loop through the files in GG-Data so that it can run all of them rather than manually typing in the year.